Rename tote_bag ui directory to be more specific
The GUI code in there is all to be used with JUCE.
So let's be explicit at the directory level.
