+++ libs/tote_bag/juce_gui/components/panels/VerticalMeterPanel.cpp
@@ -10,7 +10,7 @@
 
 #include "VerticalMeterPanel.h"
 
-#include "tote_bag/gui/lookandfeel/LookAndFeelConstants.h"
+#include "tote_bag/juce_gui/lookandfeel/LookAndFeelConstants.h"
 
 #include <juce_gui_basics/juce_gui_basics.h>
