+++ libs/tote_bag/juce_gui/components/panels/VerticalMeterPanel.h
@@ -10,7 +10,7 @@
 
 #pragma once
 
-#include "tote_bag/gui/lookandfeel/LookAndFeel.h"
+#include "tote_bag/juce_gui/lookandfeel/LookAndFeel.h"
 
 #include <juce_gui_basics/juce_gui_basics.h>
