+++ libs/tote_bag/juce_gui/components/widgets/LabelSlider.cpp
