+++ libs/tote_bag/juce_gui/components/widgets/LabelSlider.h
@@ -10,7 +10,7 @@
 
 #pragma once
 
-#include "tote_bag/gui/components/widgets/ParameterSlider.h"
+#include "tote_bag/juce_gui/components/widgets/ParameterSlider.h"
 
 #include <juce_gui_basics/juce_gui_basics.h>
