+++ libs/tote_bag/juce_gui/lookandfeel/MeterLookAndFeelMethods.h
