+++ libs/tote_bag/juce_gui/managers/ToteBagPresetManager.h
