+++ libs/tote_bag/juce_gui/utilities/GraphicsUtilities.cpp
